Refactor character creation to use new controller -> router flow [Re: nmkataoka/adv-life#69]
Created the Router to separate out external events (basically requests)
from internal events. This commit refactors the character-creation
controller to use Router instead of the EventSys.
