+++ .eslintrc
@@ -22,7 +22,7 @@
     "import/extensions": ["off"],
     "indent": 0, // using @typescript-eslint/indent
     "max-classes-per-file": 0,
-    "max-len": ["warn", { "code": 120 }],
+    "max-len": 0, // disabled because we have prettier
     "no-mixed-operators": 0,
     "no-param-reassign": 0,
     "no-plusplus": 0,
