+++ src/0-engine/API/Controller.ts
@@ -1,6 +1,8 @@
+import { Router } from './Router';
+
 /* Handles external interactions */
 export abstract class Controller {
-  public Start(): void {}
+  public abstract Start(router: Router): void;
 }
 
 export interface ControllerConstructor<C extends Controller> {
